requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
appdirs==1.4.4
asgiref==3.3.1; python_version >= '3.5'
black==20.8b1
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-environ==0.4.5
django==3.1.7
djangorestframework==3.12.2
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pytz==2021.1
regex==2021.3.17
sqlparse==0.4.1; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typed-ast==1.4.2
typing-extensions==3.7.4.3
